Adding clean code1
