+++ main/pages/HomePage/selectors/HomePage.selectors.js
@@ -1,0 +1,3 @@
+export const applicationLink = `div[class="card mt-4 top-card"]`;
+export const practiseFormButton = `div[class="element-list collapse show"]`;
+export const formButton = `div[class="card mt-4 top-card"]`;
