+++ main/pages/productDetailsPage/selectors/productDetailsPage.selectors.js
@@ -1,0 +1,1 @@
+export const details = `div[class*="col-md-9 col-sm-12"]`;
