+++ main/pages/productsPage/selectors/productsPage.selectors.js
@@ -1,0 +1,3 @@
+export const learningJavascriptBook = `a[href="/books?book=9781449331818"]`;
+export const searchButton = `span[id="basic-addon2"]`;
+export const searchField = `input[id="searchBox"]`;
