--- pages/Homepage/actions/clickApplicationLink.js
@@ -1,8 +1,0 @@
-import * as selectors from "../../../selectors/myAllSelectors.js";
-
-export async function clickApplicationLink() {
-  let link = await $$(selectors.applicationLink);
-  await link[5].scrollIntoView();
-  await link[5].waitForDisplayed();
-  await link[5].click();
-}
