+++ tests/features/toolsQA.feature
@@ -7,13 +7,13 @@ When the user clicks the book store application
 When the user searches for book from the search bar
 When the user clicks the book
 Then the user verifies that correct book details are shown on the webpage
-|Headings     | inputs|
-|ISBN        |      9781449331818|   
-|Title        |"Learning JavaScript Design Patterns"|
+|Headings     | inputs                                    |
+|ISBN         |      9781449331818                        |   
+|Title        |"Learning JavaScript Design Patterns"      |
 |sub_Title    |"A JavaScript and jQuery Developer's Guide"|
-|Author       |"Addy Osmani"|
-|Publisher    |"O'Reilly Media" |
-|Total_Pages |254       | 
+|Author       |"Addy Osmani"                              |
+|Publisher    |"O'Reilly Media"                           |
+|Total_Pages  |254                                        | 
 
 
 Scenario:The practice form is filled correctly by the given data
@@ -22,6 +22,6 @@ When the user clicks the forms button
 When the user selects practise form
 When the text data is filled by the user
 When the file of pic is entered by the user
-# # # Then the format of file is verified by the user
+Then the format of file is verified by the user
