+++ wdio.conf.js
@@ -16,7 +16,7 @@ exports.config = {
   // NPM script (see https://docs.npmjs.com/cli/run-script) then the current working
   // directory is where your package.json resides, so `wdio` will be called from there.
   //
-  specs: ["./tests/Project1/features/toolsQA.feature"],
+  specs: ["./tests/features/toolsQA.feature"],
   // Patterns to exclude.
   exclude: [
     // 'path/to/excluded/files'
@@ -124,7 +124,7 @@ exports.config = {
   //
   // If you are using Cucumber you need to specify the location of your step definitions.
   cucumberOpts: {
-    require: ["./tests/Project1/stepDefinitions/toolsQAStep.js"], // <string[]> (file/dir) require files before executing features
+    require: ["./tests/stepDefinitions/toolsQAStep.js"], // <string[]> (file/dir) require files before executing features
     backtrace: false, // <boolean> show full backtrace for errors
     requireModule: [], // <string[]> ("extension:module") require files with the given EXTENSION after requiring MODULE (repeatable)
     dryRun: false, // <boolean> invoke formatters without executing steps
